LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2-expected.txt

Button1  Button2  Invoker1  Button3  Button4
Invoker  after
Show popover
Toggle popover Popover with focusable element Other focusable element

FAIL Popover focus navigation assert_equals: Hidden popover should be skipped expected Element node <button id="button2">Button2</button> but got Element node <button id="button1">Button1</button>
FAIL Circular reference tab navigation assert_equals: Step 2 expected Element node <button id="circular1" autofocus="" popovertarget="popove... but got Element node <button id="circular0" popovertarget="popover4">Invoker</...
FAIL Popover focus returns when popover is hidden by invoker assert_true: popover should be invoked by invoker expected true got false
FAIL Popover focus only returns to invoker when focus is within the popover assert_equals: next up is the popover expected Element node <button>focusable element</button> but got Element node <button popovertarget="focus-return2-p">Toggle popover</b...
